chore: extract zip functions to separate file
