+++ src/node_dependencies/index.js
@@ -1,5 +1,6 @@
 const { dirname, basename, normalize } = require('path')
 
+const findUp = require('find-up')
 const { not: notJunk } = require('junk')
 const precinct = require('precinct')
 
@@ -14,6 +15,10 @@ const {
 } = require('./traverse')
 const { getTreeFiles } = require('./tree_files')
 const { shouldTreeShake } = require('./tree_shake')
+
+const AUTO_PLUGINS_DIR = '.netlify/plugins/'
+
+const getPluginsModulesPath = (srcDir) => findUp(`${AUTO_PLUGINS_DIR}node_modules`, { cwd: srcDir, type: 'directory' })
 
 // Retrieve the paths to the Node.js files to zip.
 // We only include the files actually needed by the function because AWS Lambda
@@ -119,5 +124,6 @@ module.exports = {
   getDependencyNamesAndPathsForDependencies,
   getDependencyNamesAndPathsForDependency,
   getExternalAndIgnoredModulesFromSpecialCases,
+  getPluginsModulesPath,
   listFilesUsingLegacyBundler,
 }
